Split check env read tests
