+++ docs/PHASE_PLAN.md
@@ -493,7 +493,7 @@ checked-in docs, tests, and commits only.
   fallback accept/reject cases in a focused child module, leaving env-effect
   ordering coverage in the parent integration module.
 - Check-command build graph host-effect tests now also keep env-read declared
-  fallback and rejection-ordering coverage in a focused child module, leaving
+  fallback and rejection-ordering coverage in focused child modules, leaving
   the parent integration module as host-effect module wiring only.
 - Build-command host-effect tests now keep declared file-read fallback accept
   and reject cases in a focused child module, leaving env-effect and
