+++ tests/integration/cli_build/graph_validation_host_effects.rs
@@ -1,4 +1,4 @@
-#[path = "graph_validation_host_effects/env_reads.rs"]
+#[path = "graph_validation_host_effects/env_reads/mod.rs"]
 mod env_reads;
 #[path = "graph_validation_host_effects/file_reads/mod.rs"]
 mod file_reads;
