+++ tests/integration/cli_build/graph_validation_host_effects/env_reads/mod.rs
@@ -1,0 +1,4 @@
+#[path = "declared.rs"]
+mod declared;
+#[path = "rejections.rs"]
+mod rejections;
